[#998] 시스템 설정 열기를 Routing 공통 계약으로 올린다
- Routing — openSystemSetting 을 요구사항 + extension 기본 구현으로 승격. BaseRouterImple 을 상속하지 않는 ApplicationRootRouter 도 함께 얻는다
- AIAgentRouter/CalendarViewRouterImple/DayEventListRouter/EventSettingRouter — 각자 복제하던 구현 제거하고 상속분 사용
- EventNotificationDefaultTimeOptionRouting/SelectEventNotificationTimeRouting — 이름만 달랐던 openSystemNotificationSetting 을 같은 계약으로 흡수
- BaseSpyRouter — didOpenSystemSetting 기록을 공통 더블로 흡수
