+++ Supports/TestDoubles/Sources/Scenes/BaseSpyRouter.swift
@@ -66,6 +66,11 @@ open class BaseSpyRouter: Routing {
         self.didShowWebViewPath = path
     }
 
+    public var didOpenSystemSetting: Bool?
+    open func openSystemSetting() {
+        self.didOpenSystemSetting = true
+    }
+
     public var didDismissPresented: Bool?
     public func dismissPresented(animated: Bool, _ completed: (() -> Void)?) {
